mapa_bolivia/requirements.txt
Flask==2.3.3
opencv-python-headless==4.9.0.80
numpy==1.26.4
python-dotenv==1.0.1
